# temp_openapi.json
{"openapi":"3.1.0","info":{"title":"Devin's Nest API 接口文档","version":"0.1.0"},"paths":{"/api/v1/ai/chat":{"post":{"tags":["ai"],"summary":"Chat Completion","description":"Generate a chat completion using the configured LLM.","operationId":"chat_completion_api_v1_ai_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/nest/projects/":{"get":{"tags":["nest-projects"],"summary":"Read Projects","description":"Retrieve projects for the current user (DevinNest Web).","operationId":"read_projects_api_v1_nest_projects__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"},"title":"Response Read Projects Api V1 Nest Projects  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["nest-projects"],"summary":"Create Project","description":"Create a new project.","operationId":"create_project_api_v1_nest_projects__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/auth/login":{"post":{"tags":["backstage-auth"],"summary":"用户登录","description":"用户名密码登录，获取 Token。","operationId":"login_api_v1_backstage_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/auth/me":{"get":{"tags":["backstage-auth"],"summary":"获取当前用户信息","description":"校验 Token 有效性并获取用户信息。\nIn a real app, you would decode the token to get the user ID.","operationId":"get_current_user_info_api_v1_backstage_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfo"}}}}}}},"/api/v1/backstage/auth/logout":{"post":{"tags":["backstage-auth"],"summary":"Logout","description":"Logout (Client side clears token).","operationId":"logout_api_v1_backstage_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/backstage/dashboard/stats":{"get":{"tags":["backstage-dashboard"],"summary":"获取统计数据","description":"获取首页所需的各项统计指标。","operationId":"get_dashboard_stats_api_v1_backstage_dashboard_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardStats"}}}}}}},"/api/v1/backstage/blogs":{"get":{"tags":["backstage-blogs"],"summary":"获取博客列表","description":"获取博客列表，支持分页和筛选。","operationId":"get_blogs_api_v1_backstage_blogs_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Pagesize"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"keyword","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Keyword"}},{"name":"categoryId","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Categoryid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["backstage-blogs"],"summary":"Create Blog","description":"Create a new blog.","operationId":"create_blog_api_v1_backstage_blogs_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Blog"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/blogs/{id}":{"get":{"tags":["backstage-blogs"],"summary":"获取博客详情","description":"获取博客详情。","operationId":"get_blog_detail_api_v1_backstage_blogs__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Blog"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["backstage-blogs"],"summary":"更新博客","description":"更新博客文章。","operationId":"update_blog_api_v1_backstage_blogs__id__put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Blog"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["backstage-blogs"],"summary":"Delete Blog","description":"Delete blog.","operationId":"delete_blog_api_v1_backstage_blogs__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/snippets":{"get":{"tags":["backstage-snippets"],"summary":"获取碎片列表","operationId":"get_snippets_api_v1_backstage_snippets_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Pagesize"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Snippet"},"title":"Response Get Snippets Api V1 Backstage Snippets Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["backstage-snippets"],"summary":"Create Snippet","operationId":"create_snippet_api_v1_backstage_snippets_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SnippetCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Snippet"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/snippets/{id}":{"get":{"tags":["backstage-snippets"],"summary":"获取碎片详情","operationId":"get_snippet_detail_api_v1_backstage_snippets__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Snippet"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["backstage-snippets"],"summary":"Update Snippet","operationId":"update_snippet_api_v1_backstage_snippets__id__put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SnippetCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Snippet"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["backstage-snippets"],"summary":"删除碎片","operationId":"delete_snippet_api_v1_backstage_snippets__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/categories":{"get":{"tags":["backstage-taxonomy"],"summary":"获取分类列表","operationId":"get_categories_api_v1_backstage_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Category"},"type":"array","title":"Response Get Categories Api V1 Backstage Categories Get"}}}}}},"post":{"tags":["backstage-taxonomy"],"summary":"创建分类","operationId":"create_category_api_v1_backstage_categories_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/categories/{id}":{"delete":{"tags":["backstage-taxonomy"],"summary":"删除分类","operationId":"delete_category_api_v1_backstage_categories__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/tags":{"get":{"tags":["backstage-taxonomy"],"summary":"获取标签列表","operationId":"get_tags_api_v1_backstage_tags_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Tag"},"type":"array","title":"Response Get Tags Api V1 Backstage Tags Get"}}}}}},"post":{"tags":["backstage-taxonomy"],"summary":"创建标签","operationId":"create_tag_api_v1_backstage_tags_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/tags/{id}":{"delete":{"tags":["backstage-taxonomy"],"summary":"删除标签","operationId":"delete_tag_api_v1_backstage_tags__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backstage/upload":{"post":{"tags":["backstage-common"],"summary":"文件上传","description":"上传文件到服务器。","operationId":"upload_file_api_v1_backstage_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_api_v1_backstage_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Blog":{"properties":{"title":{"type":"string","title":"Title"},"subtitle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subtitle"},"cover":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cover"},"categoryId":{"type":"string","title":"Categoryid"},"tagIds":{"items":{"type":"string"},"type":"array","title":"Tagids","default":[]},"status":{"type":"string","title":"Status","default":"draft"},"id":{"type":"string","title":"Id"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"views":{"type":"integer","title":"Views","default":0},"createdAt":{"type":"string","format":"date-time","title":"Createdat"},"category":{"anyOf":[{"$ref":"#/components/schemas/Category"},{"type":"null"}]},"tags":{"items":{"$ref":"#/components/schemas/Tag"},"type":"array","title":"Tags","default":[]}},"type":"object","required":["title","categoryId","id","createdAt"],"title":"Blog"},"BlogCreate":{"properties":{"title":{"type":"string","title":"Title"},"subtitle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subtitle"},"cover":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cover"},"categoryId":{"type":"string","title":"Categoryid"},"tagIds":{"items":{"type":"string"},"type":"array","title":"Tagids","default":[]},"status":{"type":"string","title":"Status","default":"draft"},"content":{"type":"string","title":"Content"}},"type":"object","required":["title","categoryId","content"],"title":"BlogCreate"},"BlogListResponse":{"properties":{"list":{"items":{"$ref":"#/components/schemas/Blog"},"type":"array","title":"List"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["list","total"],"title":"BlogListResponse"},"BlogUpdate":{"properties":{"title":{"type":"string","title":"Title"},"subtitle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subtitle"},"cover":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cover"},"categoryId":{"type":"string","title":"Categoryid"},"tagIds":{"items":{"type":"string"},"type":"array","title":"Tagids","default":[]},"status":{"type":"string","title":"Status","default":"draft"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"}},"type":"object","required":["title","categoryId"],"title":"BlogUpdate"},"Body_upload_file_api_v1_backstage_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_api_v1_backstage_upload_post"},"Category":{"properties":{"name":{"type":"string","title":"Name"},"icon":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Icon"},"color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Color"},"id":{"type":"string","title":"Id"},"count":{"type":"integer","title":"Count","default":0}},"type":"object","required":["name","id"],"title":"Category"},"CategoryCreate":{"properties":{"name":{"type":"string","title":"Name"},"icon":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Icon"},"color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Color"}},"type":"object","required":["name"],"title":"CategoryCreate"},"ChatMessage":{"properties":{"role":{"type":"string","pattern":"^(user|assistant|system)$","title":"Role"},"content":{"type":"string","title":"Content"}},"type":"object","required":["role","content"],"title":"ChatMessage"},"ChatRequest":{"properties":{"messages":{"items":{"$ref":"#/components/schemas/ChatMessage"},"type":"array","title":"Messages"},"model":{"type":"string","title":"Model","default":"gpt-3.5-turbo"},"temperature":{"type":"number","maximum":2.0,"minimum":0.0,"title":"Temperature","default":0.7}},"type":"object","required":["messages"],"title":"ChatRequest"},"ChatResponse":{"properties":{"id":{"type":"string","title":"Id"},"choices":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Choices"},"created":{"type":"integer","title":"Created"},"model":{"type":"string","title":"Model"}},"type":"object","required":["id","choices","created","model"],"title":"ChatResponse"},"DashboardStats":{"properties":{"blogsCount":{"type":"integer","title":"Blogscount"},"snippetsCount":{"type":"integer","title":"Snippetscount"},"categoriesCount":{"type":"integer","title":"Categoriescount"},"tagsCount":{"type":"integer","title":"Tagscount"},"latestActivity":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Latestactivity"}},"type":"object","required":["blogsCount","snippetsCount","categoriesCount","tagsCount"],"title":"DashboardStats"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"LoginResponse":{"properties":{"token":{"type":"string","title":"Token"},"userInfo":{"$ref":"#/components/schemas/UserInfo"}},"type":"object","required":["token","userInfo"],"title":"LoginResponse"},"Project":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"owner":{"type":"string","title":"Owner"}},"type":"object","required":["id","name","description","owner"],"title":"Project"},"Snippet":{"properties":{"content":{"items":{"$ref":"#/components/schemas/SnippetBlock"},"type":"array","title":"Content"},"metadata":{"$ref":"#/components/schemas/SnippetMetadata"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"id":{"type":"string","title":"Id"}},"type":"object","required":["content","metadata","id"],"title":"Snippet"},"SnippetBlock":{"properties":{"type":{"type":"string","title":"Type"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"src":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Src"},"caption":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Caption"},"author":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author"}},"type":"object","required":["type"],"title":"SnippetBlock"},"SnippetCreate":{"properties":{"content":{"items":{"$ref":"#/components/schemas/SnippetBlock"},"type":"array","title":"Content"},"metadata":{"$ref":"#/components/schemas/SnippetMetadata"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]}},"type":"object","required":["content","metadata"],"title":"SnippetCreate"},"SnippetMetadata":{"properties":{"weather":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Weather"},"mood":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mood"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"date":{"type":"string","format":"date-time","title":"Date"}},"type":"object","required":["date"],"title":"SnippetMetadata"},"Tag":{"properties":{"name":{"type":"string","title":"Name"},"color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Color"},"id":{"type":"string","title":"Id"},"count":{"type":"integer","title":"Count","default":0}},"type":"object","required":["name","id"],"title":"Tag"},"TagCreate":{"properties":{"name":{"type":"string","title":"Name"},"color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Color"}},"type":"object","required":["name"],"title":"TagCreate"},"UploadResponse":{"properties":{"url":{"type":"string","title":"Url"},"filename":{"type":"string","title":"Filename"}},"type":"object","required":["url","filename"],"title":"UploadResponse"},"UserInfo":{"properties":{"id":{"type":"string","title":"Id"},"username":{"type":"string","title":"Username"},"avatar":{"type":"string","title":"Avatar"}},"type":"object","required":["id","username","avatar"],"title":"UserInfo"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}